+++ packages/rollup-plugin/test/options.test.ts
@@ -1,11 +1,11 @@
 import { expect } from 'chai'
-import { 
-  MAX_RETRIES, 
-  DEFAULT_ENDPOINT, 
-  DEFAULT_REVISION, 
+import {
+  MAX_RETRIES,
+  DEFAULT_ENDPOINT,
+  DEFAULT_REVISION,
   DEFAULT_SILENT,
   DEPLOY_ENDPOINT,
-  cleanOptions, 
+  cleanOptions,
 } from '../src/options';
 
 describe('Options', () => {
@@ -14,31 +14,40 @@ describe('Options', () => {
       expect(cleanOptions.bind(cleanOptions, {})).to.throw('apiKey is required')
     });
 
-    it('should not allow retries above the MAX_RETRIES', () => {
-      const result = cleanOptions({ 
-        apiKey: 'test_key', 
+    it('should error if ignorePaths is not an array', () => {
+      expect(cleanOptions.bind(cleanOptions, {
+        apiKey: 'test_key',
         assetsUrl: 'https://foo.bar',
-        retries: 100 
+        ignorePaths: 'foo'
+      })).to.throw('ignorePaths must be an array')
+    });
+
+    it('should not allow retries above the MAX_RETRIES', () => {
+      const result = cleanOptions({
+        apiKey: 'test_key',
+        assetsUrl: 'https://foo.bar',
+        retries: 100
       })
       expect(result.retries).to.equal(MAX_RETRIES)
     })
 
     it('should merge in default options', () => {
-      const result = cleanOptions({ 
-        apiKey: 'test_key', 
+      const result = cleanOptions({
+        apiKey: 'test_key',
         assetsUrl: 'https://foo.bar',
-        retries: 0, 
+        retries: 0,
         deploy: { localUsername: 'BethanyBerkowitz' },
       })
       expect(result).to.deep.equal({
-        apiKey: 'test_key', 
+        apiKey: 'test_key',
         assetsUrl: 'https://foo.bar',
-        retries: 0, 
-        endpoint: DEFAULT_ENDPOINT, 
-        revision: DEFAULT_REVISION, 
-        silent: DEFAULT_SILENT, 
+        retries: 0,
+        endpoint: DEFAULT_ENDPOINT,
+        revision: DEFAULT_REVISION,
+        silent: DEFAULT_SILENT,
         deploy: { localUsername: 'BethanyBerkowitz' },
         deployEndpoint: DEPLOY_ENDPOINT,
+        ignorePaths: [],
       })
     })
   });