+++ packages/rollup-plugin/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.0](https://github.com/honeybadger-io/honeybadger-js/compare/v5.4.1...v5.5.0) (2023-05-11)
+
+
+### Features
+
+* **rollup-plugin:** rely on the sourcemap and js files being in the same folder and named the same ([#1063](https://github.com/honeybadger-io/honeybadger-js/issues/1063)) ([0ea16b6](https://github.com/honeybadger-io/honeybadger-js/commit/0ea16b64f6fa901856f6a9bfdce4ba00c9db6db2))
+
+
+
+## [5.4.1](https://github.com/honeybadger-io/honeybadger-js/compare/v5.4.0...v5.4.1) (2023-05-08)
+
+
+### Bug Fixes
+
+* **rollup-plugin:** document rollup peer dependency ([#1061](https://github.com/honeybadger-io/honeybadger-js/issues/1061)) ([d64a681](https://github.com/honeybadger-io/honeybadger-js/commit/d64a6819e7d0a43a576387d6dd18739edd1236e2))
+
+
+
+## [5.4.0](https://github.com/honeybadger-io/honeybadger-js/compare/v5.3.2...v5.4.0) (2023-04-27)
+
+
+### Features
+
+* add ignorePaths option ([#1059](https://github.com/honeybadger-io/honeybadger-js/issues/1059)) ([27c357e](https://github.com/honeybadger-io/honeybadger-js/commit/27c357ee02adc4f49062c4b6283c2dc3faf57546))
+
+
+
 ## [5.1.5](https://github.com/honeybadger-io/honeybadger-js/compare/v5.1.4...v5.1.5) (2023-03-06)
 
 **Note:** Version bump only for package @honeybadger-io/rollup-plugin